Proposed changes for French translation
This changes are from one of our members on skymatixva.com Andy G, who have done several contributions on this and other small changes for the site.

He propose changes based on context text validation, to make sure it does make sense. This is a second iteration of changes, which seems to be final, but further proposed changes will be submitted.